% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backdec.R
\name{backdec}
\alias{backdec}
\title{Inverse function to wholebinary returns the number from its expansion}
\usage{
backdec(rmat, sgn)
}
\arguments{
\item{rmat}{is a matrix of two rows, the first row
of the matrix is the expansion of the integer part
the second row is the binary expansion of the fractional part.}

\item{sgn}{is the sign}
}
\description{
Inverse function to wholebinary returns the number from its expansion
}
\note{
It may be necessary to make a new version of this
using special functions for large integers.
}
