% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borelmerge.R
\name{borelmerge}
\alias{borelmerge}
\title{Auxiliary function that takes avector and produces a single number through 
a Borel isomorphism using the wholebinary and backdec functions.}
\usage{
borelmerge(xvec)
}
\arguments{
\item{xvec}{is a vector of real numbers}
}
\value{
produces a single real number by converting each element
}
\description{
Auxiliary function that takes avector and produces a single number through 
a Borel isomorphism using the wholebinary and backdec functions.
}
