% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracbinary.R
\name{fracbinary}
\alias{fracbinary}
\title{Take fractionary part and make its binary expansion
Auxiliary function used in expanding real numbers}
\usage{
fracbinary(x)
}
\arguments{
\item{x}{is a number between 0 and 1}
}
\value{
Binary expansion of length 31 of the decimal input
}
\description{
Take fractionary part and make its binary expansion
Auxiliary function used in expanding real numbers
}
\note{
this implementation uses the built-in function intToBits
}
