% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbinary.R
\name{numbinary}
\alias{numbinary}
\title{Computes the binary expansion of a number}
\usage{
numbinary(x)
}
\arguments{
\item{x}{is a real or integer number}
}
\value{
the output is a binary vector of length 31
}
\description{
If the argument x is a real number the decimal
portion is dropped.
}
