% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{weave}
\alias{weave}
\title{Take a matrix of two numbers given in their binary expansion
one in each of the two rows
and return the interleaving of the two numbers}
\usage{
weave(rmat, sgn)
}
\arguments{
\item{rmat}{a matrix with two times m rows corresponding to the 
the expansions of the m numbers to be interleaved.}

\item{sgn}{is the sign vector associated to the numbers to be weaved}
}
\description{
Take a matrix of two numbers given in their binary expansion
one in each of the two rows
and return the interleaving of the two numbers
}
