% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wholebinary.R
\name{wholebinary}
\alias{wholebinary}
\title{Encodes a number as a two row binary matrix and its sign}
\usage{
wholebinary(x, nc = 31)
}
\arguments{
\item{x}{is a decimal number}

\item{nc}{is the length of the binary expansion and defines the
number of columns of the output matrix}
}
\value{
This function generates a list with a binary matrix
rmat
with two rows and the sign sgn in a separate entry of the list.
}
\description{
Auxiliary function used for generating expansion of a number,
the binary expansion of length nc of the integer part is the first row 
and the binary expansion of length nc of the fractional part
is the second row of the matrix.
The sign as appended into the final list object which the function
returns.
}
