% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xicor.ci.R
\name{xicor.ci}
\alias{xicor.ci}
\title{Computes confidence intervals for rank correlation coefficient xi.}
\usage{
xicor.ci(x, y, conf = 0.95, method = "all", R = 2000)
}
\arguments{
\item{x}{Vector of numeric values in the first coordinate.}

\item{y}{Vector of numeric values in the second coordinate.}

\item{conf}{The confidence level.}

\item{method}{The method for constructing the interval; can be any combination of
"boot.norm", "boot.nonpar", "sigma.lh", or "all" for all methods.(see Description).}

\item{R}{The number of bootstrap repetitions.}
}
\value{
A data frame with columns:
  \item{method}{The method used for the confidence interval}
  \item{lower}{The lower bound of the confidence interval}
  \item{upper}{The upper bound of the confidence interval}

  The confidence level is stored as an attribute "conf".
}
\description{
This function implements different methods for computing a confidence
interval for xicor. As the common n-out-of-n bootstrap has been shown
to fail both for estimating the variance [1] and for estimating the
distribution of xicor [3], bootstrap methods utilize the m-out-of-n
bootstrap. The parameter `methods` can be any combination of the
following strings, or "all" for all methods:
\describe{\item{"boot.norm"}{An interval based on asymptotic normality
and root-n consistency of xicor. The standard deviation is estimated
with an n-choose-m bootstrap. Suggested by Dette & Kroll [2].}
\item{"boot.nonpar"}{A non-parametric n-choose-m bootstrap interval
based on root-n consistency of xicor. The bootstrap is done with a
normalized version of xicor that has lower bias. Suggested by
Dalitz, Arning, & Goebbels [3].}
\item{"sigma.lh"}{An interval based on asymptotic normality for
continuous x and y. The standard deviation is estimated with the
second estimator (overlined sigma) by Lin & Han [4]. As this estimator
can yield negative values for the variance, especially for small sample
sizes, computation of a confidence interval with this method can fail.}
}
}
\examples{

x <- runif(100, min=-1, max=1)
y <- x^2 + rnorm(100, sd=0.2)
ci <- xicor.ci(x, y)
print(ci)

# Compute CI using only bootstrap normal method
ci_boot_norm <- xicor.ci(x, y, method = "boot.norm")
print(ci_boot_norm)

}
\references{
\describe{\item{[1]}{Lin, Z. and Han, F. (2024):
_On the failure of the bootstrap for Chatterjee’s rank correlation_
Biometrika, asae004, 02,
<doi:10.1093/biomet/asae004>.}
\item{[2]}{Dette, H. and Kroll, M. (2024):
_A simple bootstrap for Chatterjee’s rank correlation_
Biometrika, asae045,
<doi:10.1093/biomet/asae045>.}
\item{[3]}{Dalitz, C., Arning, J. and Goebbels, S. (2024)
_A Simple Bias Reduction for Chatterjee’s Correlation_
Journal of Statistical Theory and Practice 18, 51,
<doi:10.1007/s42519-024-00399-y>.}
\item{[4]}{Lin, H. and Han, F. (2022):
_Limit theorems of Chatterjee’s rank correlation_
<arxiv:2204.08031>.}
}
}
\seealso{
xicor
}
\author{
Christoph Dalitz, Juliane Arning, Felix Lögler
}
